code/requirements.txt
asgiref
Django==4.2.1
pytz
sqlparse
channels-redis==4.1.0
channels==3.0.2
django-filter==23.2
djangorestframework==3.14.0
numpy==1.24.3
pandas==2.0.1
protobuf
transformers==4.27.1
cpm_kernels
torch>=1.10
gradio
mdtex2html
sentencepiece
accelerate
torchvision
decouple
psycopg2==2.9.6
torchpack
timm
onnx
onnxsim
onnxruntime